Rewrote error for all tasks.
I rewrote how error is defined in all tasks. I did this to make the code read more like physics. I wanted my velocityy and position terms to represent the actual velocity and position terms (not the velocity of error). I did this because I wanted the correction to be of the form -kx - dx', just like it is in physics. Doing this meant interperting the error as the position minus the target, not the other way around. This way, the derivative of the error is the velocity.

I added some comments to explain the logic behind the new correction terms. These can be found in the Auto_Drive and Auto_Turn tasks.

Further, I ran into a number of issues with the drive assist. The corrections would only work in one direction. After some extensive debugging, I found that this was caused by a typecasting issue. For some reason, robotc would cast any negative float to 0 (as a Byte). To fix this, I specicially casted the -kx-dx' term as a signed char. This fixed the issue. I repeated this casting on the drive and turn tasks.

+++ Auto_Drive.h
@@ -16,7 +16,7 @@ task Auto_Drive();
 	#define Drive_Min_Velocity 20                             // In ticks per second
 
 task Drive_Assist();
-	short Drive_Assist_Enable = 0;
+	Byte Drive_Assist_Enable = 0;
 	Byte L_Drive_Correction = 0;
 	Byte R_Drive_Correction = 0;
 	#define P_Drive_Assist 3.5
